Adding Symmetry Breaking for Hard Cases
The case with

s+1  0   1
 1  s+1  0
 0   1  s+1

in the 3x3 matrix and all 0s in coordinates 2 and 3 of w{3,19,35,67}
(which is also the first one to be listed in the DNF file) has proven
to be much harder than the rest. These symmetry breaking clauses in
the preprocessing target breaking symmetries for that case.