Managing the Output Through a Specialized Class
This is a rather intrusive change but it will help merge the slurm
branch with master and allow testers to run the sequential or the
parallel workflows.

With these changes, the tools to execute in the toolchain are now
optional and the family of files output depends on how many tools are
passed as arguments.

This version only outputs files with the number of the sequence (so it
no longer outputs s[346].cnf). Makefile rules have been updated to
reflect this.

Tested for s=3.